Student Project Allocation and Management
ABSTRACT:

            The project “Student Project Allocation and Management” is aimed at developing a Web based system, which manages the activity of “Student Project Allocation and Management” and “Online Testing”. This system will manage the database and maintain a list of all student groups that have registered on this site, conduct their online test and shortlist those students who have passed the eligibility criteria as set by the professors.

Existing System:

           Up to now there is no extendible software availability for this process. The existing system of the Student Project Management is run manually by entering data in books and keeping track of all the data in the books. Adding, searching is quite difficult if it is it the form of long books.

So this software is well required for the Student project management Information System.

Problems in Existing System

The existing system has lot of problems such as:

The entire database is maintained manually which is rather tedious and error prone.
Time delay is more because the professor has to go through many records to see his student’s performance etc.
Objectives of Proposed system

The proposed system tries to solve the problems mentioned above. The main objective of the proposed system is to provide information instantly as and when it is required. The main objective is to make the Hostel information details more efficient and effective. This system should maintain different data files, so that the data can be retrieved easily and in an efficient manner. The system is very interactive. It should ensure process integration to the desired extent, various reports should be generated as the need be. This system should also ensure that there is no redundancy in the recorded data.

 Hardware Environment :

             CPU                                     : Pentium 4, 1.7 GHz

             RAM                                      : 512 MB

             Hard disk                              : 80 GB       

Software Environment:

            Operating Environment      : Windows200/Xp-SP2

             RDBMS Package                : Oracle

            Front End                              : JSP.

             Browser                                : IE7